requirements.txt
# Django框架 - 降级版本以支持Python3.6
Django>=3.2,<4.0
djangorestframework>=3.11.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.7.0

# 数据库
psycopg2-binary>=2.9.0
redis>=3.5.0

# 任务队列 - 进一步降级版本以兼容Python3.6
celery>=5.1.0
django-celery-beat>=2.2.0
django-celery-results>=2.2.0

# Git操作
GitPython>=3.1.0

# AI集成
requests>=2.25.0

# 配置管理
PyYAML>=5.4.0

# 密码加密
cryptography>=3.4.0

# API文档
drf-yasg>=1.20.0

# 测试
pytest>=6.2.0
pytest-django>=4.4.0
factory-boy>=3.2.0

# 配置管理
python-dotenv>=0.19.0

# 部署
gunicorn>=20.1.0
whitenoise>=5.3.0

# ========== 会议助手功能依赖 ==========

# WebSocket支持 - 降级版本以支持Python3.6
channels==4.0.0
channels-redis==4.2.1

# ASR服务（可选）
# funasr>=1.0.0

# NLP服务
jieba

# LLM服务（可选）
# ollama>=0.1.0

# 文档导出
markdown2>=2.4.0
# weasyprint>=60.0  # PDF导出（需要系统依赖）
python-docx>=1.1.0  # Word导出

# ========== 知识库文档处理依赖 ==========
PyPDF2>=3.0.0  # PDF处理
beautifulsoup4>=4.9.0  # HTML/Markdown解析

# ========== RAG 知识库依赖 (可选) ==========
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.21.0
